app-extension: make token table common
