+++ packages/app-extension/src/components/Unlocked/Balances/TransferWidget.tsx
@@ -1,46 +1,14 @@
 import { Blockchain } from "@coral-xyz/common";
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { Typography } from "@mui/material";
-import { styles, useCustomTheme } from "@coral-xyz/themes";
+import { useCustomTheme } from "@coral-xyz/themes";
 import { ArrowUpward, ArrowDownward } from "@mui/icons-material";
-import {
-  TextField,
-  BalancesTable,
-  BalancesTableHead,
-  BalancesTableContent,
-  BalancesTableRow,
-  BalancesTableCell,
-} from "@coral-xyz/react-xnft-renderer";
-import {
-  useBlockchainLogo,
-  useBlockchainTokensSorted,
-} from "@coral-xyz/recoil";
 import { WithHeaderButton } from "./TokensWidget/Token";
 import { Deposit } from "./TokensWidget/Deposit";
 import { Send } from "./TokensWidget/Send";
 import { useNavStack } from "../../Layout/NavStack";
-
-type Token = ReturnType<typeof useBlockchainTokensSorted>[number];
-
-const useStyles = styles((theme) => ({
-  searchField: {
-    marginLeft: "12px",
-    marginRight: "12px",
-    marginTop: "16px",
-    marginBottom: "16px",
-    width: "inherit",
-    display: "flex",
-    "& .MuiOutlinedInput-root": {
-      height: "48px !important",
-      "& fieldset": {
-        border: `solid 2pt ${theme.custom.colors.border}`,
-      },
-      "&:hover fieldset": {
-        border: `solid 2pt ${theme.custom.colors.primaryButton}`,
-      },
-    },
-  },
-}));
+import type { Token } from "../../common/TokenTable";
+import { SearchableTokenTable } from "../../common/TokenTable";
 
 export function TransferWidget() {
   return (
@@ -164,88 +132,17 @@ function TransferButton({
 }
 
 function SendToken() {
-  const classes = useStyles();
-  const [searchFilter, setSearchFilter] = useState("");
-  return (
-    <div>
-      <TextField
-        placeholder={"Search"}
-        value={searchFilter}
-        setValue={setSearchFilter}
-        rootClass={classes.searchField}
-        inputProps={{
-          style: {
-            height: "48px",
-          },
-        }}
-      />
-      <TokenTable searchFilter={searchFilter} />
-    </div>
-  );
-}
-
-function TokenTable({ searchFilter }: { searchFilter?: string }) {
-  const blockchain = Blockchain.SOLANA;
-  const title = "Tokens";
-
-  const blockchainLogo = useBlockchainLogo(blockchain);
-  const tokenAccountsSorted = useBlockchainTokensSorted(blockchain);
-  const [search, setSearch] = useState(searchFilter ?? "");
-
-  const searchLower = search.toLowerCase();
-  const tokenAccountsFiltered = tokenAccountsSorted.filter(
-    (t) =>
-      t.nativeBalance !== 0 &&
-      t.name &&
-      (t.name.toLowerCase().startsWith(searchLower) ||
-        t.ticker.toLowerCase().startsWith(searchLower))
-  );
-
-  useEffect(() => {
-    setSearch(searchFilter ?? "");
-  }, [searchFilter]);
-
-  return (
-    <BalancesTable>
-      <BalancesTableHead
-        props={{ title, iconUrl: blockchainLogo, disableToggle: true }}
-      />
-      <BalancesTableContent>
-        {tokenAccountsFiltered.map((token) => (
-          <TokenRow key={token.address} token={token} blockchain={blockchain} />
-        ))}
-      </BalancesTableContent>
-    </BalancesTable>
-  );
-}
-
-function TokenRow({
-  token,
-  blockchain,
-}: {
-  token: Token;
-  blockchain: Blockchain;
-}) {
   const { push } = useNavStack();
+
+  const onClickRow = (blockchain: Blockchain, token: Token) => {
+    push("send", { blockchain, token });
+  };
+
   return (
-    <BalancesTableRow
-      onClick={() =>
-        push("send", {
-          blockchain,
-          token,
-        })
-      }
-    >
-      <BalancesTableCell
-        props={{
-          icon: token.logo,
-          title: token.name,
-          subtitle: `${token.nativeBalance.toLocaleString()} ${token.ticker}`,
-          usdValue: token.usdBalance,
-          percentChange: token.recentUsdBalanceChange,
-        }}
-      />
-    </BalancesTableRow>
+    <SearchableTokenTable
+      onClickRow={onClickRow}
+      customFilter={(token: Token) => token.nativeBalance !== 0}
+    />
   );
 }
